fix(jsx): remove unnecessary container component
